+++ src/test/resources/projects/compilationlevel/expected/advanced.js
@@ -1,2 +1,1 @@
-'use strict';function b(a){return 1>a?1:a*b(a-1)}const c={g:99,i:b(20),h:2,l:3};console.log(c.g);console.log(c.j);console.log(c.h);
-
+'use strict';function b(a){return a<1?1:a*b(a-1)}const c={g:99,i:b(20),h:2,l:3};console.log(c.g);console.log(c.j);console.log(c.h);